requirements.txt
requests
feedparser
python-dateutil
python-dotenv
yfinance==0.2.26
openai
beautifulsoup4